Implemented softDeletes and timestamps for creating and updating

+++ src/Models/Model.php
@@ -5,9 +5,13 @@ namespace Src\Models;
 use Exception;
 use Src\Database\Connection;
 use PDO;
+use DateTime;
+use Src\Traits\HasTimestamps;
+use Src\Traits\HasSoftDeletes;
 
 class Model
 {
+    use HasTimestamps, HasSoftDeletes;
     protected $db;
     protected $pdo;
     protected $table;
@@ -24,12 +28,18 @@ class Model
 
     public function save()
     {
+        if ($this->timestampsEnabled) {
+            $this->setTimestamps($this->attributes);
+        }
+
         if (!isset($this->id)) {
             $this->id = $this->db->insert($this->table, $this->attributes);
             $this->attributes[$this->primaryKeyName] = $this->id;
+            echo "Item saved with ID: " . $this->id . PHP_EOL;
         } else {
             $this->db->update($this->table, $this->attributes, [$this->primaryKeyName => $this->id]);
         }
+
     }
 
     public function update(string $id)
@@ -37,7 +47,14 @@ class Model
         if (is_null($this->id)) {
             throw new Exception('Cannot update record without primary key being set what ID it should use.');
         }
+
+        if ($this->timestampsEnabled) {
+            $now = (new DateTime())->format('Y-m-d H:i:s');
+            $this->attributes['updated_at'] = $now;
+        }
+
         $this->db->update($this->table, $this->attributes, [$this->primaryKeyName => $this->id]);
+        echo 'Item with ID ' . $this->id . 'updated.' . PHP_EOL;
     }
 
     public static function find(string $primaryKeyValue)
@@ -48,7 +65,6 @@ class Model
         $result = $db->fetchAssoc($query, ['primaryKey' => $primaryKeyValue]);
 
         if ($result) {
-            echo 'The requested ID was found in the database.' . PHP_EOL;
             $instance->attributes = $result;
             $instance->id = $result[$instance->primaryKeyName];
             return $instance;
@@ -59,18 +75,23 @@ class Model
 
     public function delete(string $id)
     {
-        $this->id= $id;
+        $this->id = $id;
         if (is_null($this->id)) {
             throw new Exception("Cannot delete record without a primary key value.");
         }
 
-        $query = "DELETE FROM {$this->table} WHERE {$this->primaryKeyName} = :primaryKey LIMIT 1";
-        $stmt = $this->pdo->prepare($query);
-        $stmt->bindParam(':primaryKey', $this->id);
-        $stmt->execute();
-        echo 'Deletion sucessful.' . PHP_EOL;
-        if ($stmt->rowCount() > 0) {
-            return true;
+        if ($this->softDeletesEnabled) {
+            $this->softDelete($id);
+        } else {
+            $query = "DELETE FROM {$this->table} WHERE {$this->primaryKeyName} = :primaryKey LIMIT 1";
+            $stmt = $this->pdo->prepare($query);
+            $stmt->bindParam(':primaryKey', $this->id);
+            $stmt->execute();
+            echo 'Item with ID ' . $this->id . ' deleted.' . PHP_EOL;
+            if ($stmt->rowCount() > 0) {
+                return true;
+            }
+
         }
 
         return false;
